fix: address review comments for httproute plugin refs
- Remove unnecessary ensureKubeconfig function
- Extract index functions to httproute_index.go
- Simplify key collection with slices.Collect(maps.Keys())
- Move referenceGrants loading outside plugin loop
- Scope listReferenceGrants to relevant namespace
- Extract errInvalidType sentinel error
